code-assistant/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.4.0
sentence-transformers==2.2.2
GitPython==3.1.40
python-dotenv==1.0.0
